geoevolve/requirements.txt
openai>=1.3.0
chromadb>=0.4.0
pykrige>=1.7.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
matplotlib>=3.8.0

